merger/requirements.txt
python-dotenv
jsonschema
requests